requirements.txt
Flask
tensorflow
numpy
google-generativeai